fix: BUG-2/3/4/5 추가 수정
- BUG-2: question.api.js review()에서 action 필드를 status로 변경 (백엔드 스키마 호환)
- BUG-3: questionRepository.upsertTypeStats() $5 파라미터 타입 불일치 해결 (::integer 명시 캐스팅)
- BUG-4: QAPage sendMessage()에서 user.academyId 대신 session.academy_id/teacher_id 사용
- BUG-5: examService.generateStudentExam()에서 OpenAI 401 에러를 503으로 변환해 프론트 로그아웃 방지

+++ back/src/domains/question/questionRepository.js
@@ -110,11 +110,11 @@ export const submitAnswer = async ({ student_id, question_id, academy_id, submit
 export const upsertTypeStats = async ({ student_id, academy_id, type_code, subject_id, is_correct }) => {
   await pool.query(
     `INSERT INTO student_type_stats (student_id, academy_id, type_code, subject_id, total_attempts, correct_count, correct_rate, last_attempted_at)
-     VALUES ($1, $2, $3, $4, 1, $5, $5, NOW())
+     VALUES ($1, $2, $3, $4, 1, $5::integer, $5::decimal, NOW())
      ON CONFLICT (student_id, type_code) DO UPDATE SET
        total_attempts = student_type_stats.total_attempts + 1,
-       correct_count = student_type_stats.correct_count + $5,
-       correct_rate = (student_type_stats.correct_count + $5)::decimal / (student_type_stats.total_attempts + 1),
+       correct_count = student_type_stats.correct_count + $5::integer,
+       correct_rate = (student_type_stats.correct_count + $5::integer)::decimal / (student_type_stats.total_attempts + 1),
        last_attempted_at = NOW(),
        updated_at = NOW()`,
     [student_id, academy_id, type_code, subject_id, is_correct ? 1 : 0]
